flask/requirements.txt
Flask==3.0.3
mysql-connector-python==9.1.0
python-dotenv==1.0.1
Flask-Cors==5.0.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0